+++ Changelog.md
@@ -6,6 +6,7 @@
 ##### 2023-XX-YY
 
 - Added the `description` and `state` fields to the master product variants table
+- Added the Address comparison utility
 - Fixed the `Carrier::name()` method to overcome accessor infinite loop
 - Fixed the address creation with order factory when passing an explicit address type